EventWatcher: reject waitingFor promise during test cleanup.
EventWatcher installs a test cleanup function that removes its event
listeners. For this reason, if the EventWatcher's wait_for created a
promise that is still pending when the test ends, the promise will never
be resolved or rejected.

This omission turns into a problem when the EventWatcher's wait_for
promise ends up in the promise chain used by promise_test. If an
assertion fails in a step() callback, the test's done() method is
called, which cleans up the EventWatcher's listener. So, a failure in a
single promise_test can block the whole promise chain and prevent all
subsequent tests from running.

This commit fixes the issue described above by having any pending EventWatcher
promise get rejected when the event listeners get removed.