Fail in the presence of duplicate test names
Test names are presumed to be unique within test files--this allows
consumers to use them for identification purposes. Duplicated names
violate this expectation and should therefore be rejected. Previously,
the default "reporter" implementation identified this case only
informally (through modification to the test document)--the error was
not communicated to whatever "runner" process initiated test execution.

Formally report an error in the presence of duplicate test names. Remove
the equivalent code in the default "reporter" implementation (as the
error is now communicated via the typical "test failure" mechanism).