Require "thenable" for `promise_test`
The `promise_test` function is intended to run an asynchronous test,
interpreting its status via a "thenable" object provided via return
value. Previously, it would accept functions that did not provide such a
value, and this meant that tests could fail silently. Because there is
no use case for using `promise_test` in this way, the condition always
reflects a programming error.

Update the harness to explicitly fail the test whenever this occurs,
prompting contributors to correct their usage prior to submitting the
test.